edited humidity reading to an integer value

+++ db/migrate/20161110055721_create_humidity_sensor_readings.rb
@@ -1,7 +1,7 @@
 class CreateHumiditySensorReadings < ActiveRecord::Migration[5.0]
   def change
     create_table :humidity_sensor_readings do |t|
-      t.float :measurement
+      t.integer :measurement
 
       t.timestamps
     end
